The Update for 16 Sept 2016
The registration section for the seller has been modified with the step
2 section with name and password filds and if the user put his phone
number in the step 1 he will we redirected to the step 3 section with
the opt option.

The loging section for the seller has been created . Now the seller can
loging using the username and password they used during the time of
registration.

After the login the seller lands to the new profile  page where he can
update his filds.The new profile page section design has been
intigrated.

The seller can can update his profile with all the details. eg
(Description , Availability )

The fild for the profile has been validate for the user input and the
data is stored into the database.